+++ IcVibracoes.Core/Models/BeamCharacteristics/FasteningType.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using IcVibracoes.DataContracts;
+using System;
 
 namespace IcVibracoes.Core.Models.BeamCharacteristics
 {
@@ -47,28 +48,45 @@ namespace IcVibracoes.Core.Models.BeamCharacteristics
     }
 
     /// <summary>
+    /// The fastenings that can be used in the analysis.
+    /// </summary>
+    public enum Fastenings
+    {
+        None = 0,
+        Pinned = 1,
+        Fixed = 2,
+    }
+
+    /// <summary>
     /// It's responsible to create a fastening object based on a string.
     /// </summary>
     public class FasteningFactory
     {
-        public static FasteningType Create(string fastening)
+        /// <summary>
+        /// This method creates an instance of class <seealso cref="FasteningType"/>.
+        /// It can be <seealso cref="None"/>, <seealso cref="Fixed"/> or <seealso cref="Pinned"/>.
+        /// </summary>
+        /// <typeparam name="TResponseData"></typeparam>
+        /// <param name="fastening"></param>
+        /// <param name="response"></param>
+        /// <returns></returns>
+        public static FasteningType Create<TResponseData>(string fastening, OperationResponseBase<TResponseData> response)
+            where TResponseData : OperationResponseData
         {
             switch ((Fastenings)Enum.Parse(typeof(Fastenings), fastening, ignoreCase: true))
             {
-                case Fastenings.Fixed: return new Fixed();
-                case Fastenings.Pinned: return new Pinned();
-                case Fastenings.None: return new None();
-                default: break;
+                case Fastenings.Fixed: 
+                    return new Fixed();
+                case Fastenings.Pinned: 
+                    return new Pinned();
+                case Fastenings.None: 
+                    return new None();
+                default: 
+                    break;
             }
 
-            throw new Exception($"Invalid fastening: {fastening}.");
+            response.AddError(OperationErrorCode.InternalServerError, $"Invalid fastening: '{fastening}'.");
+            return null;
         }
-    }
-
-    public enum Fastenings
-    {
-        None,
-        Pinned,
-        Fixed,
     }
 }