+++ IcVibracoes.Core/Operations/CalculateVibration/FiniteElement/BeamWithDva/CalculateBeamWithDvaVibration.cs
@@ -140,7 +140,8 @@ namespace IcVibracoes.Core.Operations.CalculateVibration.FiniteElement.BeamWithD
             double[] forces = beam.Forces;
 
             // Creating input.
-            FiniteElementMethodInput input = new FiniteElementMethodInput(NumericalMethodFactory.Create(request.NumericalMethod))
+            var numericalMethod = (NumericalMethod)Enum.Parse(typeof(NumericalMethod), request.NumericalMethod, ignoreCase: true);
+            FiniteElementMethodInput input = new FiniteElementMethodInput(numericalMethod)
             {
                 Mass = await this._boundaryCondition.Apply(massWithDva, bondaryCondition, numberOfTrueBoundaryConditions + (uint)beam.DvaNodePositions.Length),